requirements.txt
pdf2image>=3.1.0
opencv-python>=4.8.0
Pillow>=10.0.0
openai>=1.0.0
psutil>=5.9.0
python-dotenv>=1.0.0
numpy>=1.24.0 